1. Added "main:app" and reload=True 2. Added the creation of the src/public/bouquet folder to setup.bat 3. Removed order routes for orders, renamed the BouquetFile route to UploadFile 4. Added comments to the code. 5. Translated everything into English. 6. Added a class config for each model in which the json_schema example was specified. 7. Added app.bat, which starts the server after checking all modules and launches the log console, which displays information from the terminal window. 8. Added config_file, which stores global variables; the file is recreated with the variables specified inside after each application launch. 9. Moved a huge amount of code to services and added #comments for large services. 10. Fixed order creation; now, when a bouquet in quantity 0 is sent, an error is displayed. (you cannot create a zero order). 11. Fixed order creation; now, when the ID of a non-existent bouquet is passed in the bouquet sheet that is passed for the order, an error is displayed. (if at least 1 of the bouquets in the sheet does not exist). 12. Added try except to each route.

+++ src/service/bouquet_services.py
@@ -1,5 +1,10 @@
-from src.repository import bouquet_repository
+import os
+from fastapi import HTTPException
+from src.repository import bouquet_repository, bouquet_repository
+from src.service import image_services
 from src.database.models import Bouquet
+from src.utils.custom_logging import setup_logging
+log = setup_logging()
 
 
 async def get_all_bouquets():
@@ -28,5 +33,22 @@ async def update_bouquet(bouquet_id: int, bouquet: Bouquet):
 
 
 async def delete_bouquet(bouquet_id: int):
+    # Проверка существования букета
+    bouquet = await get_bouquet_by_id(bouquet_id)
+    if not bouquet:
+        raise HTTPException(status_code=404, detail="Bouquet not exist")
+    # Проверка существования информации о пути изображения букета
+    image = await image_services.get_image_by_id(bouquet.ImageID)
+    if not image:
+        raise HTTPException(status_code=404, detail="Image not exist")
+    # Удаление изображение букета
+    if os.path.exists(image.Url):
+        os.remove(image.Url)
+    else:
+        log.warning(f"Warning: File {image.Url} not found")
+    # Удаляем сам букет
     await bouquet_repository.delete_bouquet(bouquet_id)
+    # Удаляем записи о букете в таблице с путями изображений
+    await image_services.delete_image(image.ID)
+    log.info(f"Bouquet {bouquet_id} deleted successfully")
     return {"message": "Bouquet deleted successfully"}
